% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtAlongTrackDistance}
\alias{dtAlongTrackDistance}
\title{Along Track Distance}
\usage{
dtAlongTrackDistance(latFrom, lonFrom, latTo, lonTo, pointLat, pointLon,
  tolerance = 1e+09)
}
\arguments{
\item{latFrom}{latitude value for the start of the path}

\item{lonFrom}{longitude value for the sttart of the path}

\item{latTo}{lattitude value for the end of the path}

\item{lonTo}{longitude value for the end of the path}

\item{pointLat}{latitude value for the point}

\item{pointLon}{longitude value for the point}

\item{tolerance}{}
}
\description{
Calculates the distance along a great-circle to the cross-track point
}
\seealso{
dtDist2gc
}
