% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtBearing}
\alias{dtBearing}
\title{dt bearing}
\usage{
dtBearing(latFrom, lonFrom, latTo, lonTo, compassBearing = FALSE)
}
\arguments{
\item{latFrom}{latitude from}

\item{lonFrom}{longitude from}

\item{latTo}{latitude to}

\item{lonTo}{longitude to}

\item{compassBearing}{logical indicating whether to return the value in the range 0 - 360}
}
\value{
bearing in degrees
}
\description{
Calculates the initial bearing between two pairs of latitude / longitude coordinates
}
\note{
The initial bearing is sometimes called the 'forward azimuth', which if followed in a straight line along a great-circle arc will take you from teh start point to the end point
}
\examples{
dtBearing(0,0,0,52)

dtBearing(25, 0, 35, 0)

dtBearing(25, 0, -25, 0)

dtBearing(0, 0, 0, -1)
dtBearing(0, 0, 0, -1, compassBearing = TRUE)

dtBearing(0, 0, 1, -1)

dtBearing(0, 0, 0, 1)

dtBearing(0, 0, 1, -0.001)
dtBearing(0, 0, 1, -0.001, compassBearing = TRUE)

dtBearing(-37,144,-38,145)

}
