% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtCosine}
\alias{dtCosine}
\title{dt cosine}
\usage{
dtCosine(latFrom, lonFrom, latTo, lonTo)
}
\arguments{
\item{latFrom}{latitude}

\item{lonFrom}{longitude}

\item{latTo}{latitude}

\item{lonTo}{latitude}
}
\value{
distance in metres
}
\description{
Calculates the Cosine distance between two points
}
\examples{
dt <- data.table(lat1 = seq(-38, -37, by = 0.1),
  lon1 = seq(144, 145, by = 0.1),
  lat2 = seq(-35, -34, by = 0.1),
  lon2 = seq(145, 146, by = 0.1))

dt[, distance := dtCosine(lat1, lon1, lat2, lon2)]

}
