% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtDestination}
\alias{dtDestination}
\title{dt destination}
\usage{
dtDestination(latFrom, lonFrom, distance, bearing)
}
\arguments{
\item{latFrom}{latitude from (in degrees)}

\item{lonFrom}{longitude from (in degrees)}

\item{distance}{distance in metres}

\item{bearing}{from north (in degrees)}
}
\value{
list of latitude and longitude coordinates at the destination
}
\description{
Calculates the destination coordinates from a starting point, initial bearing and distance travelling along a (shortest distance) great circle arc
}
\examples{

dtDestination(0, 0, earthsRadius(), 90)

n <- 10
set.seed(20170511)
lats <- -90:90
lons <- -180:180
b <- 0:360
dt <- data.table::data.table(lat = sample(lats, size = n, replace = T),
                             lon = sample(lons, size = n, replace = T),
                             bearing = sample(b, size = n, replace = T))
dt[, c("destinationLat", "destinationLon") := dtDestination(lat, lon, 90)]

}
