% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtDist2gc}
\alias{dtDist2gc}
\title{Dist 2 gc}
\usage{
dtDist2gc(latFrom, lonFrom, latTo, lonTo, pointLat, pointLon,
  tolerance = 1e+09, r = earthsRadius())
}
\arguments{
\item{latFrom}{latitude value for the start of the path}

\item{lonFrom}{longitude value for the sttart of the path}

\item{latTo}{lattitude value for the end of the path}

\item{lonTo}{longitude value for the end of the path}

\item{pointLat}{latitude value for the point}

\item{pointLon}{longitude value for the point}

\item{r}{}
}
\value{
distance in metres. The sign of the result indicates if the point
is to the left (negative) or right (positive) of the line when traveling
from `latFrom,lonFrom` to `latTo,lonTo`.
}
\description{
Finds the distance of a point to a great-circle path
}
\details{
Sometimes called 'cross-track' distance
}
\examples{
dt <- data.table(lat1 = c(0, 0),
                 lon1 = c(0, 0),
                 lat2 = c(20, -20),
                 lon2 = c(0, 0),
                 plat = c(10, -10),
                 plon = c(22, 22))

dt[, dist := dtDist2gc(lat1, lon1, lat2, lon2, plat, plon)]

}
