% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtEuclidean}
\alias{dtEuclidean}
\title{dt Euclidean}
\usage{
dtEuclidean(latFrom, lonFrom, latTo, lonTo)
}
\arguments{
\item{latFrom}{latitude from}

\item{lonFrom}{longitude from}

\item{latTo}{latitude to}

\item{lonTo}{longitude to}
}
\value{
Euclidean distance
}
\description{
Computes the Euclidean distance between two pairs of latitude / longitude coordinates
}
\examples{
dt <- data.table(lat1 = seq(-38, -37, by = 0.1),
  lon1 = seq(144, 145, by = 0.1),
  lat2 = seq(-35, -34, by = 0.1),
  lon2 = seq(145, 146, by = 0.1))

dt[, distance := dtEuclidean(lat1, lon1, lat2, lon2)]

}
