% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtHaversine}
\alias{dtHaversine}
\title{dt haversine}
\usage{
dtHaversine(latFrom, lonFrom, latTo, lonTo, tolerance = 1e+09)
}
\arguments{
\item{latFrom}{latitude}

\item{lonFrom}{longitude}

\item{latTo}{latitude}

\item{lonTo}{latitude}

\item{tolerance}{numeric. See details}
}
\value{
distance in metres
}
\description{
Calculates the Haversine distance between two points
}
\details{
\code{tolerance} - Floating-point precision can sometimes lead to the
situation `sqrt(1 - 1.00000001)`, which will introduce NAs into the result.

The tolerance value is used in `ifelse(a > 1 & a <= tolerance, 1, a)`
}
\examples{
dt <- data.table(lat1 = seq(-38, -37, by = 0.1),
  lon1 = seq(144, 145, by = 0.1),
  lat2 = seq(-35, -34, by = 0.1),
  lon2 = seq(145, 146, by = 0.1))

dt[, distance := dtHaversine(lat1, lon1, lat2, lon2)]

}
