% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceCalculations.R
\name{dtMidpoint}
\alias{dtMidpoint}
\title{dt midpoint}
\usage{
dtMidpoint(latFrom, lonFrom, latTo, lonTo)
}
\arguments{
\item{latFrom}{latitude from}

\item{lonFrom}{longitude from}

\item{latTo}{latitude to}

\item{lonTo}{longitude to}
}
\value{
list of the latitude & longitude coordinates at the midpoint along the line
}
\description{
Finds the half-way point along the great circle path between two points
}
\examples{
dtMidpoint(0,0,0,52)

dtMidpoint(25, 0, -25, 0)

dtMidpoint(-37,144,-38,145)

dtMidpoint(25, 0, 35, 0)

n <- 10
set.seed(20170511)
lats <- -90:90
lons <- -180:180
dt <- data.table::data.table(lat1 = sample(lats, size = n, replace = T),
                             lon1 = sample(lons, size = n, replace = T),
                             lat2 = sample(lats, size = n, replace = T),
                             lon2 = sample(lons, size = n, replace = T))

dt[, c("latMid", "lonMid") := dtMidpoint(lat1, lon1, lat2, lon2)]


}
