% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_lat_column}
\alias{find_lat_column}
\title{Find Lat Column}
\usage{
find_lat_column(names, msg, stopOnFailure = TRUE)
}
\description{
Attempts to find the column of a data.frame containing the latitude values
}
