% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_lon_column}
\alias{find_lon_column}
\title{Find Lon Column}
\usage{
find_lon_column(names, msg, stopOnFailure = TRUE)
}
\description{
Attempts to find the column of a data.frame containing the longitude values
}
