% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialJoins.R
\name{isLeft}
\alias{isLeft}
\title{Is Left}
\usage{
isLeft(p0x, p0y, p1x, p1y, p2x, p2y)
}
\description{
Tests if a point is Left|On|Right of an infinite line
Returns:
 > 0 : P2 is left of the line through P0, P1
 = 0 : P2 is on the line
 < 0 : P2 is right of the line through P0, P1
}
