% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialJoins.R
\name{PointInPolygon}
\alias{PointInPolygon}
\title{Point In Polygon}
\usage{
PointInPolygon(dt_polygons, polyColumns, dt_points, pointColumns)
}
\arguments{
\item{dt_polygons}{\code{data.table} object containing the polygon coordinates}

\item{polyColumns}{character vector of column names containing the id, lineId, x, y and hole fields (in that order)}

\item{dt_points}{\code{data.table} object containing the point coordinates}

\item{pointColumns}{character vector of column names containing the id, x and y fields (in that order)}
}
\value{
\code{data.table} giving the ids of the points and the polygons within which they fall
}
\description{
Determins if points lie within the boundaries of polygons
}
\details{
Calculates the winding number (\url{https://en.wikipedia.org/wiki/Winding_number})
}
\examples{
\dontrun{
 ## example
}
}
