% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonCalculations.R
\name{PointsInPolygon}
\alias{PointsInPolygon}
\title{Points In Polygon}
\usage{
PointsInPolygon(vectorX, vectorY, pointsX, pointsY, pointsIds)
}
\arguments{
\item{pointsY}{}
}
\description{
Determins if a set of points fall within a polygon
}
\details{
Calculates the winding number (\url{https://en.wikipedia.org/wiki/Winding_number})
}
\examples{
\dontrun{
 ## example
}
}
