% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline.R
\name{PolylineDistance}
\alias{PolylineDistance}
\title{Polyline Distance}
\usage{
PolylineDistance(polyline)
}
\arguments{
\item{polyline}{}
}
\description{
Calculates the distance of an encoded polyline
}
\examples{
\dontrun{

sdt <- copy(spdt_melbourne)

sdt[, polyline_distance := PolylineDistance(polyline)]


}

}
