% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdatatable.R
\name{polylines}
\alias{polylines}
\title{Polylines}
\usage{
polylines(sdt)
}
\arguments{
\item{sdt}{spatial.data.table object}
}
\description{
gets the encoded polyline(s) from a spatial.data.table object
}
