% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_tramRoute}
\alias{sf_tramRoute}
\title{sf Tram Route}
\format{A \code{sf} object of 2 observations and 1 variable
\describe{
  \item{geometry}{sfc_LINESTRING of length 2}
}}
\usage{
sf_tramRoute
}
\description{
Simple Features (\code{library(sf)}) object of the route 35 Tram in Melbourne
}
\keyword{datasets}
