% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{spatialdatatable-package}
\alias{spatialdatatable}
\alias{spatialdatatable-package}
\title{spatial.data.table}
\description{
A \code{spatial.data.table} is an extension of \code{data.table}. All the usual
operations on a \code{data.table} should still work on a \code{spatial.data.table}.
}
\author{
\strong{Maintainer}: David Cooley \email{dcooley@symbolix.com.au}

}
