% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeSF.R
\name{EncodeSF}
\alias{EncodeSF}
\title{Encode Simple Feature}
\usage{
EncodeSF(sf)
}
\arguments{
\item{sf}{simple-feature object (from \code{library(sf)})}
}
\description{
Converts coordinates from simple-feature objects (\code{sf}) into encoded polylines
}
\examples{
library(sf)

sf_poly <- st_as_sfc(c("POLYGON((-80.190 25.774, -66.118 18.466, -64.757 32.321))",
                       "POLYGON((-70.579 28.745, -67.514 29.570, -66.668 27.339))"))
sf <- st_sf(id = paste0("poly", 1:2), sf_poly)

EncodeSF(sf)

}
