% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialdatatable.R
\name{polyline_column}
\alias{polyline_column}
\title{Polyline Column}
\usage{
polyline_column(sdt)
}
\arguments{
\item{sdt}{spatialdatatable object}
}
\description{
Gets the column(s) names of a spatialdatatable object containing encoded polylines
}
