% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline.R
\name{PolylineArea}
\alias{PolylineArea}
\title{Polyline Area}
\usage{
PolylineArea(polyline)
}
\arguments{
\item{polyline}{}
}
\description{
Calculates the area of an enclosed polyline.
}
\examples{
\dontrun{

}

}
