% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline.R
\name{PolylineCenter}
\alias{PolylineCenter}
\title{Polyline Center}
\usage{
PolylineCenter(polyline)
}
\arguments{
\item{polyline}{an encoded polyline}
}
\description{
Finds the center coordinates of a polyline
}
