% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonCalculations.R
\name{WindingNumber}
\alias{WindingNumber}
\title{Winding Number}
\usage{
WindingNumber(pointX, pointY, polyX, polyY)
}
\arguments{
\item{pointX}{x-coordinate of a point (typically 'latitude')}

\item{pointY}{y-coordinate of a point (typically 'longitude')}

\item{polyX}{x-coordinates of the polygon}

\item{polyY}{y-coordinates of thepolygon}
}
\description{
Calculates the winding number for a point and polygon. Returns 0 when the point
is outside the polygon.
}
